had trouble with git on the train but just added the pictures
