--- src/day09/assets/day09ListNode.java
@@ -1,12 +1,0 @@
-package day09.assets;
-
-public class day09ListNode {
-    public int data;
-    public day09LinkedList value;
-    public day09ListNode next;
-
-    public day09ListNode(int value){
-        this.data=value;
-        this.next=null;
-    }
-}
