+++ src/day09/day09ListNode.java
@@ -1,0 +1,12 @@
+package day09;
+
+public class day09ListNode {
+    public int data;
+    public day09LinkedList value;
+    public day09ListNode next;
+
+    public day09ListNode(int value){
+        this.data=value;
+        this.next=null;
+    }
+}
